refactor: change historyList to historyScroll
